+++ src/taoensso/timbre/tools/logging.clj
@@ -1,6 +1,7 @@
 (ns taoensso.timbre.tools.logging
   "clojure.tools.logging.impl/Logger implementation"
-  (:require [taoensso.timbre :as timbre]))
+  (:require [clojure.tools.logging]
+            [taoensso.timbre :as timbre]))
 
 (deftype Logger [logger-ns]
   clojure.tools.logging.impl/Logger
@@ -9,8 +10,8 @@
     ;; tools.logging message may be a string (for `logp`/`logf` calls) or raw
     ;; argument (for `log` calls). Note that without an :args equivalent for
     ;; `write!`, the best we can do is `[message]`. This inconsistency means
-    ;; that :args consumers (like the rate limiter and Postal appender) will
-    ;; necessarily behave differently under tools.logging.
+    ;; that :args consumers will necessarily behave differently under
+    ;; tools.logging.
     (timbre/send-to-appenders! level {} [message] logger-ns throwable
       (when (string? message) message))))