+++ src/taoensso/timbre/utils.clj
@@ -10,20 +10,61 @@
   (let [[name [expr]] (macro/name-with-attributes name sigs)]
     `(clojure.core/defonce ~name ~expr)))
 
-(defn memoize-ttl
-  "Like `memoize` but invalidates the cache for a set of arguments after TTL
-  msecs has elapsed."
+(defn memoize-ttl "Low-overhead, common-case `memoize*`."
   [ttl-ms f]
   (let [cache (atom {})]
     (fn [& args]
-      (let [{:keys [time-cached d-result]} (@cache args)
-            now (System/currentTimeMillis)]
+      (when (<= (rand) 0.001) ; GC
+        (let [instant (System/currentTimeMillis)]
+          (swap! cache
+            (fn [m] (reduce-kv (fn [m* k [dv udt :as cv]]
+                                (if (> (- instant udt) ttl-ms) m*
+                                    (assoc m* k cv))) {} m)))))
+      (let [[dv udt] (@cache args)]
+        (if (and dv (< (- (System/currentTimeMillis) udt) ttl-ms)) @dv
+          (locking cache ; For thread racing
+            (let [[dv udt] (@cache args)] ; Retry after lock acquisition!
+              (if (and dv (< (- (System/currentTimeMillis) udt) ttl-ms)) @dv
+                (let [dv (delay (apply f args))
+                      cv [dv (System/currentTimeMillis)]]
+                  (swap! cache assoc args cv)
+                  @dv)))))))))
 
-        (if (and time-cached (< (- now time-cached) ttl-ms))
-          @d-result
-          (let [d-result (delay (apply f args))]
-            (swap! cache assoc args {:time-cached now :d-result d-result})
-            @d-result))))))
+(defn rate-limiter
+  "Returns a `(fn [& [id]])` that returns either `nil` (limit okay) or number of
+  msecs until next rate limit window (rate limited)."
+  [ncalls-limit window-ms]
+  (let [state (atom [nil {}])] ; [<pull> {<id> {[udt-window-start ncalls]}}]
+    (fn [& [id]]
+
+      (when (<= (rand) 0.001) ; GC
+        (let [instant (System/currentTimeMillis)]
+          (swap! state
+            (fn [[_ m]]
+              [nil (reduce-kv
+                    (fn [m* id [udt-window-start ncalls]]
+                      (if (> (- instant udt-window-start) window-ms) m*
+                          (assoc m* id [udt-window-start ncalls]))) {} m)]))))
+
+      (->
+       (let [instant (System/currentTimeMillis)]
+         (swap! state
+           (fn [[_ m]]
+             (if-let [[udt-window-start ncalls] (m id)]
+               (if (> (- instant udt-window-start) window-ms)
+                 [nil (assoc m id [instant 1])]
+                 (if (< ncalls ncalls-limit)
+                   [nil (assoc m id [udt-window-start (inc ncalls)])]
+                   [(- (+ udt-window-start window-ms) instant) m]))
+               [nil (assoc m id [instant 1])]))))
+       (nth 0)))))
+
+(comment
+  (def rl (rate-limit 10 10000))
+  (repeatedly 10 #(rl (rand-nth [:a :b :c])))
+  (rl :a)
+  (rl :b)
+  (rl :c))
 
 (defn merge-deep-with ; From clojure.contrib.map-utils
   "Like `merge-with` but merges maps recursively, applying the given fn
@@ -45,8 +86,7 @@
 (comment (merge-deep {:a {:b {:c {:d :D :e :E}}}}
                      {:a {:b {:g :G :c {:c {:f :F}}}}}))
 
-(defn round-to
-  "Rounds argument to given number of decimal places."
+(defn round-to "Rounds argument to given number of decimal places."
   [places x]
   (if (zero? places)
     (Math/round (double x))
@@ -56,11 +96,14 @@
 (comment (round-to 0 10)
          (round-to 2 10.123))
 
-(defmacro fq-keyword
-  "Returns namespaced keyword for given name."
+(defmacro fq-keyword "Returns namespaced keyword for given name."
   [name]
-  `(if (and (keyword? ~name) (namespace ~name))
-     ~name
+  `(if (and (keyword? ~name) (namespace ~name)) ~name
      (keyword (str ~*ns*) (clojure.core/name ~name))))
 
 (comment (map #(fq-keyword %) ["foo" :foo :foo/bar]))
+
+(defmacro sometimes "Executes body with probability e/o [0,1]."
+  [probability & body]
+  `(do (assert (<= 0 ~probability 1) "Probability: 0 <= p <= 1")
+       (when (< (rand) ~probability) ~@body)))